sqlalchemy: parse_dictionary now understands more operators in mongo query documents
The client or a hook can now send a mongo query document in the lookup parameter.
Allowed operators are $or, $and, $gt, $gte, $lt, $lte and $ne.
Until now only key value pairs were allowed, which would create an AND query with
the equality operator. There was no way to create an OR query using the lookup
parameter, which is the only way for a hook to add filters to the query.

Note that a malformed lookup parameter will lead to a 400 Bad Request response, although
the error might also be caused by a hook, which modified the lookup dictionary.